Add items to manual control combo boxes (#16)
Add items to the manual control combo boxes. When you click on the triangle, options with 0, 100, and 200 appear.

Rename the action buttons at the bottom of manual control. Previously it was Play, Pause, Stop. Now make it Start, Stop, Clear. Implement the clear functionality.

Also fix logger menu bar color and change the title on AboutForm to About.